refactor(api): use shared domain error for unauthorized access to entity error
